% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarsdf.R
\name{polarsdf}
\alias{polarsdf}
\title{Polars DataFrame}
\usage{
polarsdf(df)
}
\arguments{
\item{df}{A data frame or similar object.}
}
\value{
A Polars DataFrame.
}
\description{
Polars DataFrame
}
\examples{
polarsdf(iris)
}
